年齢計算の件数を共通化
